chore: fix type

+++ WoWMarketWatcher.API/Models/DTOs/Alerts/AlertConditionDto.cs
@@ -16,6 +16,6 @@ namespace WoWMarketWatcher.API.Models.DTOs.Alerts
 
         public int AggregationTimeGranularityInHours { get; init; }
 
-        public int Threshold { get; init; }
+        public long Threshold { get; init; }
     }
 }
