+++ WoWMarketWatcher.API/Models/Requests/Alerts/CreateAlertConditionRequest.cs
@@ -18,6 +18,6 @@ namespace WoWMarketWatcher.API.Models.Requests.Alerts
         public int? AggregationTimeGranularityInHours { get; init; }
 
         [Required]
-        public int? Threshold { get; init; }
+        public long? Threshold { get; init; }
     }
 }
